requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
aiohttp>=3.9.0
asyncio>=3.4.3
fake-useragent>=1.4.0
pandas>=2.1.0
openpyxl>=3.1.0
python-dateutil>=2.8.0
schedule>=1.2.0
flask>=3.0.0
flask-cors>=4.0.0
pydantic>=2.5.0
tenacity>=8.2.0
tqdm>=4.66.0
python-dotenv>=1.0.0
newspaper3k>=0.2.8
feedparser>=6.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
